% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ms_alignment.R
\name{boxplot.ms_alignment}
\alias{boxplot.ms_alignment}
\title{Make boxplot from MS peak table.}
\usage{
\method{boxplot}{ms_alignment}(x, formula, ...)
}
\arguments{
\item{x}{A peak_table object}

\item{formula}{A formula object}

\item{...}{Additional arguments to \code{\link[graphics]{boxplot}}}
}
\description{
The function can take multiple response variables on the left hand side of the
formula (separated by \code{+}). In this case, a separate boxplot will be
produced for each response variable.
}
\concept{Visualization}
