% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peak}
\alias{find_peak}
\title{Find peak based on retention time and/or mass}
\usage{
find_peak(x, rt, mz, rt.tol = 0.01, mz.tol = 0.05)
}
\arguments{
\item{x}{An \code{msdial_alignment} object}

\item{rt}{Retention time}

\item{mz}{Quant.mass}

\item{rt.tol}{Tolerance for matching retention time}

\item{mz.tol}{Tolerance for matching Quant.mass}
}
\value{
Returns EI spectrum as a \code{data.frame}.
}
\description{
Find peak based on retention time and/or mass
}
\author{
Ethan Bass
}
