% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_spectra.R
\name{get_spectrum}
\alias{get_spectrum}
\title{Get spectrum from MSDIAL alignment object}
\usage{
get_spectrum(x, col)
}
\arguments{
\item{x}{An \code{msdial_alignment} object or matrix with rows as samples and features as columns.}

\item{col}{Index of the feature (column).}
}
\value{
Returns spectrum as a data.frame with two columns: "mz" and "intensity".
}
\description{
Get spectrum from MSDIAL alignment object
}
\author{
Ethan Bass
}
