% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_msp.R
\name{match_spectra}
\alias{match_spectra}
\title{Match spectra to msp database}
\usage{
match_spectra(x, db, ..., mc.cores = 2)
}
\arguments{
\item{x}{Spectrum, as produced by \code{\link{get_spectrum}}.}

\item{db}{MSP database as list}

\item{...}{Additional arguments to \code{match_spectrum}}

\item{mc.cores}{How many cores to use for parallel processing? Defaults to 2.}
}
\description{
Match spectra to msp database
}
\author{
Ethan Bass
}
