% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_msp.R
\name{match_spectrum}
\alias{match_spectrum}
\title{Match spectrum in MSDIAL alignment to database}
\usage{
match_spectrum(
  x,
  db,
  ...,
  col,
  ri_thresh = 100,
  spectral_weight = 0.8,
  n.results = 12,
  mc.cores = 2,
  ris
)
}
\arguments{
\item{x}{Spectrum, as produced by \code{\link{get_spectrum}}.}

\item{db}{MSP database as list}

\item{...}{Additional arguments to \code{\link[OrgMassSpecR]{SpectrumSimilarity}}}

\item{col}{Index of feature to select}

\item{ri_thresh}{Maximum difference between retention indices for a match
to be considered.}

\item{spectral_weight}{A number between 0 and 1 to specify the weight given
to spectral similarity versus retention index similarity.}

\item{n.results}{How many results to return.}

\item{mc.cores}{How many cores to use for parallel processing? Defaults to 2.}

\item{ris}{Retention indices to use}
}
\description{
Match spectrum in MSDIAL alignment to database
}
\author{
Ethan Bass
}
