% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ms_calculate_RIs}
\alias{ms_calculate_RIs}
\title{Convert retention times to retention indices in alignment object.}
\usage{
ms_calculate_RIs(x, Ris)
}
\arguments{
\item{x}{An \code{msdial_alignment} object.}

\item{Ris}{A matrix or data.frame containing retention times in column one
and retention indices in column two.}
}
\description{
Convert retention times to retention indices in alignment object.
}
