% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_msdial.R
\name{ms_call_msdial}
\alias{ms_call_msdial}
\title{Call MS-DIAL console app
For help configuring the MSDIAL console app on mac OSX or linux, please see
the \href{https://github.com/Jiung-Wen/msdial}{instructions} helpfully
compiled by \href{https://github.com/Jiung-Wen/}{Jiung-Wen Chen}.}
\usage{
ms_call_msdial(
  system,
  path_in,
  path_out,
  method,
  settings,
  p = FALSE,
  mce = FALSE
)
}
\arguments{
\item{system}{Either \code{gcms}, \code{lcmsdda}, \code{lcsdia},
\code{lcmsdia}, \code{lcimmsdda}, or \code{lcimmsdia}.}

\item{path_in}{Path to files.}

\item{path_out}{Path to output directory}

\item{method}{A method file.}

\item{settings}{Settings in lieu of a method file.}

\item{p}{Logical.}

\item{mce}{Logical}
}
\value{
Returns MSDIAL alignment.
}
\description{
Call MS-DIAL console app
For help configuring the MSDIAL console app on mac OSX or linux, please see
the \href{https://github.com/Jiung-Wen/msdial}{instructions} helpfully
compiled by \href{https://github.com/Jiung-Wen/}{Jiung-Wen Chen}.
}
