% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignment.R
\name{ms_filter_alignment}
\alias{ms_filter_alignment}
\title{Filter alignment by provided indices.}
\usage{
ms_filter_alignment(x, idx, what = c("rows", "cols"), inverse = FALSE)
}
\arguments{
\item{x}{An \code{ms_alignment} object or matrix with rows as samples and features as columns.}

\item{idx}{Indices to be retained or excluded according to the value of \code{inverse}.}

\item{what}{Which dimension to filter on. Either (\code{rows}) or columns
(\code{cols}).}

\item{inverse}{Whether to retain (default) or remove the specified columns.}
}
\description{
Filter alignment by provided indices.
}
\author{
Ethan Bass
}
