% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ms_find_peak}
\alias{ms_find_peak}
\title{Find peak based on retention time and/or mass}
\usage{
ms_find_peak(x, rt, mz, rt.tol = 0.01, mz.tol = 0.05, plot_it = TRUE)
}
\arguments{
\item{x}{An \code{ms_alignment} object}

\item{rt}{Retention time}

\item{mz}{Quant.mass}

\item{rt.tol}{Tolerance for matching retention time}

\item{mz.tol}{Tolerance for matching Quant.mass}

\item{plot_it}{Logical. Whether to plot the spectra.}
}
\value{
Returns EI spectrum as a \code{data.frame}.
}
\description{
Find peak based on retention time and/or mass
}
\author{
Ethan Bass
}
