% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{ms_mirror_plot}
\alias{ms_mirror_plot}
\alias{ms_mirror_plot.data.frame}
\alias{ms_mirror_plot.msdial_alignment}
\title{Mirror plot function}
\usage{
ms_mirror_plot(x, ...)

\method{ms_mirror_plot}{data.frame}(
  x,
  y,
  plot_labels = TRUE,
  type = c("plotly", "base"),
  scale = TRUE,
  lab_int = 0.2,
  digits = 1,
  bar_width = 1,
  match_score = TRUE,
  ...
)

\method{ms_mirror_plot}{msdial_alignment}(
  x,
  cols,
  ref,
  type = c("plotly", "base"),
  scale = TRUE,
  plot_labels = TRUE,
  lab_int = 0.2,
  digits = 1,
  bar_width = 1,
  match_score = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{msdial_alignment} object.}

\item{...}{Additional arguments}

\item{y}{Mass spectrum as data.frame with m/z values in column one and
ionization intensity in column two.}

\item{plot_labels}{Logical. Whether to label m/z values on plot.}

\item{type}{What kind of plot to produce. Either base R (\code{base}) or
plotly (\code{plotly}).}

\item{scale}{Logical. Whether to scale mass spectrum. Defaults to TRUE.}

\item{lab_int}{Labels will be plotted above the specified proportion of the
largest ion.}

\item{digits}{How many figures to include on m/z labels.}

\item{bar_width}{Width of bars.}

\item{match_score}{Logical. Whether to plot match score or not.}

\item{cols}{One or more columns in the peak table \code{tab} to plot.}

\item{ref}{A row in the matches slot corresponding to the provided column.}
}
\description{
Mirror plot function

Plot two spectra as a mirror plot.

Plot two spectra as a mirror plot.
}
