% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{ms_plot_spectrum}
\alias{ms_plot_spectrum}
\title{Plot mass spectrum of peak given by \code{col}.}
\usage{
ms_plot_spectrum(
  x,
  col,
  plot_labels = TRUE,
  lab_int = 0.2,
  title = TRUE,
  type = c("plotly", "base"),
  scale = FALSE,
  bar_width = 1,
  digits = 1,
  ...
)
}
\arguments{
\item{x}{An alignment object.}

\item{col}{Spectrum to plot.}

\item{plot_labels}{Logical. Whether to plot labels or not.}

\item{lab_int}{Labels will be plotted above the specified proportion of the
largest ion.}

\item{title}{Logical. Whether to plot title. Defaults to TRUE.}

\item{type}{What kind of plot to produce. Either base R (\code{base}) or
plotly (\code{plotly})}

\item{scale}{Logical. Whether to scale mass spectrum. Defaults to FALSE.}

\item{bar_width}{Width of bars.}

\item{digits}{How many figures to include on mz labels}

\item{...}{Additional arguments.}
}
\value{
If \code{export} is \code{TRUE}, returns spectrum as \code{data.frame}.
Otherwise, no return value.
}
\description{
Plot mass spectrum of peak given by \code{col}.
}
\author{
Ethan Bass
}
