% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignment.R
\name{ms_read_alignment}
\alias{ms_read_alignment}
\title{Read MSDIAL alignment file}
\usage{
ms_read_alignment(path, format = c("msdial"))
}
\arguments{
\item{path}{Path to mass spectrometry alignment file.}

\item{format}{The format of the provided alignment file. Currently, only
MS-DIAL '.txt' files are supported (\code{msdial}).}
}
\value{
Returns \code{ms_alignment} object. A list of 3 data.frames,
containing peak data (\code{tab}), peak metadata (\code{peak_meta}) and
sample metadata (\code{sample_meta}).
}
\description{
Read MSDIAL alignment file
}
\author{
Ethan Bass
}
