% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_peaktable.R
\name{ms_reshape_peaktable}
\alias{ms_reshape_peaktable}
\title{Reshape peak table}
\usage{
ms_reshape_peaktable(
  x,
  peaks,
  metadata,
  treatments = NULL,
  fixed_levels = TRUE
)
}
\arguments{
\item{x}{An MS dial alignment object.}

\item{peaks}{A character vector specifying the peaks to include in tidy output.
If the character vector is named, the names of the vector elements will be
used in place of the original peak names.}

\item{metadata}{A character vector specifying the metadata to include in
the tidy output.}

\item{treatments}{This argument is deprecated as of version 0.3.2. It is
synonymous with the new metadata argument which should be used instead.}

\item{fixed_levels}{Fix factor levels of features in the order provided.}
}
\value{
If \code{export} is \code{TRUE}, returns spectrum as \code{data.frame}.
Otherwise, no return value.
}
\description{
Convert peak table to tidy format for plotting.
}
\author{
Ethan Bass
}
