% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_search_gadget.R
\name{ms_search_gadget}
\alias{ms_search_gadget}
\title{Launch MS search gadget for interactive viewing of spectral matches.}
\usage{
ms_search_gadget(data)
}
\arguments{
\item{data}{An \code{ms_alignment} object.}
}
\description{
Launch MS search gadget for interactive viewing of spectral matches.
}
