% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{ms_subtract_blanks}
\alias{ms_subtract_blanks}
\title{Subtract blanks}
\usage{
ms_subtract_blanks(
  x,
  blanks.idx,
  blanks.pattern,
  what = c("mean", "median"),
  drop = TRUE
)
}
\arguments{
\item{x}{A \code{ms_alignment} object.}

\item{blanks.idx}{Indices of blank samples}

\item{blanks.pattern}{A string that uniquely identifies blank samples by name}

\item{what}{Whether to subtract the mean or median value}

\item{drop}{Logical. Whether to drop columns containing only zeros. Defaults to TRUE.}
}
\value{
A \code{ms_alignment} object with the mean or median of the blanks
subtracted from each peak.
}
\description{
Subtract blanks
}
