% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ms_alignment.R
\name{plot.ms_alignment}
\alias{plot.ms_alignment}
\title{Plot spectrum from peak table}
\usage{
\method{plot}{ms_alignment}(
  x,
  col,
  plot_spectrum = TRUE,
  box_plot = FALSE,
  vars = NULL,
  spectrum_labels = TRUE,
  engine = c("base", "plotly"),
  ...
)
}
\arguments{
\item{x}{A \code{ms.alignment} object.}

\item{col}{A vector specifying the peak(s) that you wish to plot.}

\item{plot_spectrum}{Logical. If TRUE, plots the mass spectrum of the chosen
peak. Defaults to TRUE.}

\item{box_plot}{Logical. If TRUE, plots box plot using factors
defined by \code{vars}.}

\item{vars}{Independent variables for boxplot. Righthand side of formula.}

\item{spectrum_labels}{Logical. If TRUE, plots labels on maxima in spectral
plot. Defaults to TRUE.}

\item{engine}{Which plotting engine to use: either \code{base} or \code{plotly}.}

\item{...}{Additional arguments to \code{\link[graphics]{boxplot}}.}
}
\value{
No return value.
}
\description{
Plots the trace and/or spectrum for a given peak in peak table.
}
\details{
Can be used to confirm the identity of a peak or check that a particular
column in the peak table represents a single compound. Can also be used
to create simple box-plots to examine the distribution of a peak with respect
to variables defined in sample metadata.
}
\section{Side effects}{


If \code{plot_spectrum} is TRUE, plots the spectrum for the specified chromatogram
at the specified retention time. The spectrum is a single row from the chromatographic
matrix.

If \code{box_plot} is TRUE, produces a \code{\link[graphics]{boxplot}} from the
specified peak with groups provided by \code{vars}.
}

\author{
Ethan Bass
}
\concept{Visualization}
