% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot mass spectrum of peak given by \code{col}.}
\usage{
plot_spectrum(
  x,
  col,
  plot_labels = TRUE,
  lab.int = 0.2,
  type = c("plotly", "base"),
  width = 1,
  digits = 1
)
}
\arguments{
\item{x}{MS dial alignment object}

\item{col}{Spectrum to plot}

\item{plot_labels}{Logical. Whether to plot labels or not.}

\item{lab.int}{Labels will be plotted above the specified proportion of the
largest ion.}

\item{type}{What kind of plot to produce. Either base R (\code{base}) or
plotly (\code{plotly})}

\item{width}{Width of bars.}

\item{digits}{How many figures to include on mz labels}
}
\value{
No return value
}
\description{
Plot mass spectrum of peak given by \code{col}.
}
\author{
Ethan Bass
}
