% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{pqn}
\alias{pqn}
\title{Probabilistic Quotient Normalization}
\usage{
pqn(x, ref = c("median", "mean"), QC = NULL)
}
\arguments{
\item{x}{An \code{msdial_alignment} object or matrix with rows as samples and features as columns.}

\item{ref}{Reference for normalization: either \code{median} (default) to use
the overall median of variables as the reference, or \code{mean} to use the
overall average of variables as the reference.}

\item{QC}{vector of number(s) to specify samples which average to use as reference
(e.g. QC samples)}
}
\value{
A normalized \code{msdial_alignment} object or \code{matrix},
according to the input.
}
\description{
Performs Probabilistic Quotient Normalization
}
\note{
Adapted from the \href{https://github.com/ricoderks/Rcpm}{Rcpm} package
by  Rico Derks (licensed under GPL3).
}
\references{
Dieterle, F., Ross, A., Schlotterbeck, G. & Senn, H. Probabilistic Quotient
Normalization as Robust Method to Account for Dilution of Complex Biological Mixtures.
Application in H1 NMR Metabonomics. Anal. Chem. 78, 4281-4290 (2006).
}
\author{
E. Nevedomskaya

Rico Derks

Ethan Bass
}
