% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignment.R
\name{read_alignment}
\alias{read_alignment}
\title{Read MSDIAL alignment file}
\usage{
read_alignment(path)
}
\arguments{
\item{path}{Path to MS dial alignment file}
}
\value{
Returns \code{msdial_alignment} object containing peak data and metadata.
}
\description{
Read MSDIAL alignment file
}
\author{
Ethan Bass
}
