% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_spectra.R
\name{search_msp}
\alias{search_msp}
\title{Search MSP database for spectrum}
\usage{
search_msp(
  x,
  db,
  ...,
  n.results = 10,
  parallel,
  mc.cores = 2,
  what = c("msd", "df", "scores")
)
}
\arguments{
\item{x}{Spectrum, as produced by \code{\link{get_spectrum}}.}

\item{db}{MSP database as list}

\item{...}{Additional arguments to \code{\link[OrgMassSpecR]{SpectrumSimilarity}}}

\item{n.results}{Number of results to return}

\item{parallel}{Logical. Whether to use parallel processing. (This feature
does not work on Windows).}

\item{mc.cores}{How many cores to use for parallel processing? Defaults to 2.}

\item{what}{What kind of object to return. Either \code{msdial_alignment} object (\code{msd})
or \code{data.frame} (\code{df}).}
}
\description{
Search MSP database for spectrum
}
\author{
Ethan Bass
}
