% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_spectra.R
\name{spectral_similarity}
\alias{spectral_similarity}
\title{Calculate spectral similarity between two peaks
This function is slightly adapted from the \code{SpectrumSimilarity} function
in [OrgMassSpecR](https://orgmassspec.github.io/) where it is licensed under
BSD-2 (© 2011-2017, Nathan Dodder). The function was re-factored here for
increased speed.}
\usage{
spectral_similarity(
  spec.top,
  spec.bottom,
  t = 0.25,
  b = 10,
  xlim = c(50, 1200),
  x.threshold = 0
)
}
\arguments{
\item{spec.top}{data frame containing the experimental spectrum's peak list
with the m/z values in the first column and corresponding intensities in the
second.}

\item{spec.bottom}{data frame containing the reference spectrum's peak list
with the m/z values in the first column and corresponding intensities in the
second.}

\item{t}{numeric value specifying the tolerance used to align the m/z values
of the two spectra.}

\item{b}{numeric value specifying the baseline threshold for peak
identification. Expressed as a percent of the maximum intensity.}

\item{xlim}{numeric vector of length 2, defining the beginning and ending
values of the x-axis.}

\item{x.threshold}{numeric value of length 1 specifying the m/z threshold
used for the similarity score calculation. Only peaks with m/z values above
the threshold are used in the calculation. This can be used to exclude noise
and/or non-specific ions at the low end of the spectrum. By default all ions
are used.}
}
\description{
Calculate spectral similarity between two peaks
This function is slightly adapted from the \code{SpectrumSimilarity} function
in [OrgMassSpecR](https://orgmassspec.github.io/) where it is licensed under
BSD-2 (© 2011-2017, Nathan Dodder). The function was re-factored here for
increased speed.
}
\author{
Nathan G. Dodder

Ethan Bass
}
