% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{subtract_blanks}
\alias{subtract_blanks}
\title{Subtract blanks}
\usage{
subtract_blanks(x, blanks.idx, blanks.pattern, what = c("mean", "median"))
}
\arguments{
\item{x}{A \code{msdial_alignment} object.}

\item{blanks.idx}{Indices of blank samples}

\item{blanks.pattern}{A string that uniquely identifies blank samples by name}

\item{what}{Whether to subtract the mean or median value}
}
\value{
A \code{msdial_alignment} object with the mean or median of the blanks
subtracted from each peak.
}
\description{
Subtract blanks
}
