% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_eispectrum}
\alias{tidy_eispectrum}
\title{Tidy EI spectrum}
\usage{
tidy_eispectrum(x)
}
\arguments{
\item{x}{spectrum to be converted}
}
\value{
Returns EI spectrum as a \code{data.frame}.
}
\description{
Converts MS-DIAL EI spectrum to data.frame
}
\author{
Ethan Bass
}
