% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{tsn}
\alias{tsn}
\title{Total sum normalization
Divides each row by the sum of the features in that row.}
\usage{
tsn(x)
}
\arguments{
\item{x}{An \code{msdial_alignment} object or matrix with rows as samples and features as columns.}
}
\value{
A normalized \code{msdial_alignment} object or \code{matrix},
according to the input.
}
\description{
Total sum normalization
Divides each row by the sum of the features in that row.
}
