% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{ms_normalize_itsd}
\alias{ms_normalize_itsd}
\title{Normalize by internal standard}
\usage{
ms_normalize_itsd(x, idx, plot_it = FALSE)
}
\arguments{
\item{x}{An \code{msdial_alignment} object or matrix with rows as samples and features as columns.}

\item{idx}{Column index of internal standard.}

\item{plot_it}{Logical. Whether to plot ITSD against total peak area.}
}
\value{
A normalized \code{msdial_alignment} object or \code{matrix},
according to the input.
}
\description{
Normalize by internal standard
}
\author{
Ethan Bass
}
