% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ms_rt_to_ri}
\alias{ms_rt_to_ri}
\title{Convert retention times to retention indices.}
\usage{
ms_rt_to_ri(rts, RIs)
}
\arguments{
\item{rts}{A vector of retention times.}

\item{RIs}{A matrix or data.frame containing carbon number or retention index
in column one and retention times in column two.}
}
\description{
Convert retention times to retention indices.
}
