% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_spectra.R
\name{ms_search_spectra}
\alias{ms_search_spectra}
\title{Search spectra in MSDIAL alignment against database}
\usage{
ms_search_spectra(
  x,
  db,
  cols,
  ...,
  ri_thresh = 100,
  spectral_weight = 0.6,
  n_results = 10,
  parallel,
  mc.cores = 2,
  print = FALSE,
  progress_bar = TRUE
)
}
\arguments{
\item{x}{An \code{msdial_alignment} object.}

\item{db}{MSP database. The provided object should be a nested list, where the
sublists contain the following elements: retention indices in an element named
\code{RI} and mass spectra in an element called \code{Spectra}. All other elements
are optional.}

\item{cols}{Index or indices of feature(s) to be identified.}

\item{...}{Additional arguments to \code{\link[OrgMassSpecR]{SpectrumSimilarity}}.}

\item{ri_thresh}{Maximum difference between retention indices for a match.
to be considered. Defaults to 100. Use \code{NULL} to search database without
filtering by retention time (this will take a long time for large databases).}

\item{spectral_weight}{A number between 0 and 1 specifying the weight given.
to spectral similarity versus retention index similarity. Defaults to 0.6.}

\item{n_results}{How many results to return. Defaults to 10.}

\item{parallel}{Logical. Whether to use parallel processing. (This feature
does not work on Windows).}

\item{mc.cores}{How many cores to use for parallel processing? Defaults to 2.}

\item{print}{Logical. Whether to print the results after each search. Defaults
to FALSE.}

\item{progress_bar}{Logical. Whether to display progress bar or not.}
}
\value{
Returns a modified \code{msdial_alignment} object with database matches
in the \code{matches} slot as a list of data frames. Each \code{data.frame}
will contain the database matches as rows and columns corresponding to the
elements of the database entry (e.g. "Name", "InChIKey", etc.) as well as
match scores for spectral similarity (\code{spectral_match}), retention index
similarity (\code{ri_match}) and the total similarity score (\code{total_score}).
}
\description{
This function can be used to identify peaks in your peaktable by matching them
to a spectral database (\code{db}). It takes several arguments that can
be used to customize the matching algorithm, including \code{ri_thresh},
\code{spectral weight}, \code{n_results}. The retention index threshold
(\code{ri_thresh}) is used to subset the provided database, which greatly
improves the search speed. Only database entries with a retention index
falling within the specified threshold will be considered. The spectral
weight affects the weight given to spectral similarity when calculating the
the total similarity score, which is used to rank matches.
}
\note{
See \href{https://github.com/QizhiSu/mspcompiler}{mspcompiler} for help compiling
an msp database.
}
\author{
Ethan Bass
}
