% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_msdial.R
\name{ms_tidy_msdial}
\alias{ms_tidy_msdial}
\title{Convert peak to tidy format for plotting}
\usage{
ms_tidy_msdial(x, peaks, treatments)
}
\arguments{
\item{x}{An MS dial alignment object}

\item{peaks}{Peaks to include in tidy output}

\item{treatments}{Treatments to include in tidy output}
}
\value{
If \code{export} is \code{TRUE}, returns spectrum as \code{data.frame}.
Otherwise, no return value.
}
\description{
Convert peak to tidy format for plotting
}
\author{
Ethan Bass
}
