% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_itsd}
\alias{normalize_itsd}
\title{normalize by internal standard}
\usage{
normalize_itsd(x, idx)
}
\arguments{
\item{x}{An \code{msdial_alignment} object or matrix with rows as samples and features as columns.}

\item{idx}{Column index of internal standard.}
}
\description{
normalize by internal standard
}
\author{
Ethan Bass
}
