% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_spectra.R
\name{search_spectra}
\alias{search_spectra}
\title{Search spectra in MSDIAL alignment against database}
\usage{
search_spectra(
  x,
  db,
  cols,
  ...,
  ri_thresh = 100,
  spectral_weight = 0.6,
  n.results = 10,
  mc.cores = 2,
  ris
)
}
\arguments{
\item{x}{Spectrum, as produced by \code{\link{get_spectrum}}.}

\item{db}{MSP database as list}

\item{cols}{Index or indices of feature(s) to select}

\item{...}{Additional arguments to \code{\link[OrgMassSpecR]{SpectrumSimilarity}}}

\item{ri_thresh}{Maximum difference between retention indices for a match
to be considered.}

\item{spectral_weight}{A number between 0 and 1 to specify the weight given
to spectral similarity versus retention index similarity.}

\item{n.results}{How many results to return.}

\item{mc.cores}{How many cores to use for parallel processing? Defaults to 2.}

\item{ris}{Retention indices to use}
}
\description{
Search spectra in MSDIAL alignment against database
}
\author{
Ethan Bass
}
